employee-database/src/main/java/nl/com/ems/employeedatabase/exception/handler/EmployeeDataExceptionHandler.java
package nl.com.ems.employeedatabase.exception.handler;

import lombok.extern.slf4j.Slf4j;
import nl.com.ems.employeedatabase.exception.BadRequestException;
import nl.com.ems.employeedatabase.exception.InternalServerException;
import nl.com.ems.employeedatabase.exception.ResourceNotFoundException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;
import nl.com.ems.employeedatabase.model.ErrorResponse;



@RestControllerAdvice
@Slf4j
public class EmployeeDataExceptionHandler {

    @ExceptionHandler(ResourceNotFoundException.class)
    public ResponseEntity<ErrorResponse> handleResourceNotFoundException(ResourceNotFoundException ex) {
        log.error("Resource not found: {}", ex.getMessage());
        return buildErrorResponse(HttpStatus.NOT_FOUND.value(), "Resource not found", ex.getMessage());
    }

    @ExceptionHandler(BadRequestException.class)
    public ResponseEntity<ErrorResponse> handleBadRequestException(BadRequestException ex) {
        log.error("Bad request: {}", ex.getMessage());
        return buildErrorResponse(HttpStatus.BAD_REQUEST.value(), "Bad request", ex.getMessage());
    }

    @ExceptionHandler(InternalServerException.class)
    public ResponseEntity<ErrorResponse> handleInternalServerException(InternalServerException ex) {
        log.error("Internal server error: {}", ex.getMessage());
        return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Internal server error", ex.getMessage());
    }

    @ExceptionHandler(Exception.class)
    public ResponseEntity<ErrorResponse> handleException(Exception ex) {
        log.error("Unexpected error: {}", ex.getMessage());
        return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Internal server error", ex.getMessage());
    }

    private ResponseEntity<ErrorResponse> buildErrorResponse(int status, String message, String detailedMessage) {
        ErrorResponse errorResponse = new ErrorResponse();
        errorResponse.setStatus(status);
        errorResponse.setMessage(message);
        errorResponse.setDetailedMessage(detailedMessage);
        return ResponseEntity.status(status).body(errorResponse);
    }
}

